Rename log-level.ts to logging.ts for better scope representation
The file now contains both LogLevel enum and LogHandler type,
so "logging" is a more accurate name than "log-level".

Updated all import/export references accordingly.
